+++ app/src/main/java/com/example/androidkotlinsprintchallenge11/Models.kt
@@ -1,5 +1,8 @@
 package com.example.androidkotlinsprintchallenge11
 
+import kotlinx.serialization.Serializable
+
+@Serializable
 data class Model(
     val collection: String,
     val credits: String,
@@ -11,4 +14,14 @@ data class Model(
     val news_name: String,
     val short_description: String,
     val video_files: List<VideoFile>
+)
+
+@Serializable
+data class VideoFile(
+    val file_size: Int,
+    val file_url: String,
+    val format: String,
+    val frame_rate: String,
+    val height: Int,
+    val width: Int
 )
