--- app/src/main/java/com/example/androidkotlinsprintchallenge11/VideoFile.kt
@@ -1,10 +1,0 @@
-package com.example.androidkotlinsprintchallenge11
-
-data class VideoFile(
-    val file_size: Int,
-    val file_url: String,
-    val format: String,
-    val frame_rate: String,
-    val height: Int,
-    val width: Int
-)
